deleted TestNav, basic link styling for Nav, imported Nav
